fix the rare case of segment fault while uploading files
caused by the dialog automatically closing before the task is finished due to
the fact that the value is large than the maximum. and there is a bug with qt
if the task finishes too fast.

this patch adds guard to prevent it. and it also applies some human fixes for
the incorrect total_bytes during uploading (aka. not counting the network
 or http protocol cost).