Fix for #669
Fix for Client doesn't work with multiple users Mac OS #669

count_process and getBSDProcessPid where counting all processes systemwide. A check if the process was run by the current user was added. Now only processes run by the current user are considered and it is possible for each user to run one instance.