Regression fix for StatusNotifierItem protocol
The removed lines were causing a problem for the context menu to not work with qt5.6 which uses DBus, SNI and the com.canonical.dbusmenu interface. Apparently this cannot just be related to seafile but the hide() and immediately following show() call  somehow caused the Menubar object of the StatusNotifierItem to 'disappear' effectively removing the dbus  context menu object.